4. Update code in model, repo, service, controller, view pages;

+++ src/main/java/com/nesreading/domain/CartItem.java
@@ -15,7 +15,7 @@ public class CartItem {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private int id;
 
-  private long quantity;
+  private int quantity;
   private double price;
 
   @ManyToOne()
